OpenGL_Engine: Camera
git-svn-id: svn://localhost/steel.new/trunk@21 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/branches/kindex-res/code/res/model/model.h
@@ -17,7 +17,7 @@ typedef std::vector<v3> vertexes;
 // One triangle, indexes of vertexes
 struct Triangle // anti-clockwise direction (one side Triangle)
 {
-	unsigned short a[3];
+	unsigned int a[3];
 };
 
 typedef std::vector<Triangle> triangles;
